% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{person_said}
\alias{person_said}
\title{Returns a random quote / line from a character in the show}
\usage{
person_said(speaker = "Jim")
}
\arguments{
\item{speaker}{A chr}
}
\value{
If input is a character from the office, then the output is a chr vector
if input not found returns a printed mesage
}
\description{
Returns a random quote / line from a character in the show
}
\examples{
preson_said(Pam)
}
