% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{print_episode}
\alias{print_episode}
\title{Returns an html page (Rmarkdown) saved to the user's working directory}
\usage{
print_episode(season = 1, episode = 1)
}
\arguments{
\item{season}{An int}

\item{episode}{An int}
}
\value{
If input is valid episode from the show, then the output html
if input not returns null, markdown fails
}
\description{
Returns an html page (Rmarkdown) saved to the user's working directory
}
\examples{
print_episode(4, 2)
}
