% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{quote_about}
\alias{quote_about}
\title{Returns a random quote / line that that has the word parameter in it}
\usage{
quote_about(word = "nuts")
}
\arguments{
\item{word}{A chr}
}
\value{
If input is a word found in at least one quote / line from the office, then the output is a chr vector
if input not found returns a printed message
}
\description{
Returns a random quote / line that that has the word parameter in it
}
