% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{what_episode}
\alias{what_episode}
\title{Returns a random season and episode}
\usage{
what_episode()
}
\value{
the output is a chr vector
}
\description{
Returns a random season and episode
}
\examples{
what_episode()
}
